Arreglo de errores

+++ src/controller/usersManagement/resources/ResourcesControllerDelete.java
@@ -2,6 +2,7 @@ package controller.usersManagement.resources;
 
 import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
+import controller.usersManagement.access.AccessControllerView;
 import model.Resource;
 
 import javax.jdo.JDOObjectNotFoundException;
@@ -17,24 +18,37 @@ import java.io.IOException;
 @SuppressWarnings("serial")
 public class ResourcesControllerDelete extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        
-        PersistenceManager pm = controller.PMF.get().getPersistenceManager();
 
-        try {
-            Key key = KeyFactory.stringToKey(request.getParameter("key"));
-            try{
-                pm.deletePersistent(pm.getObjectById(Resource.class, key));
-                request.getSession().setAttribute("serverResponse","{\"color\": \"#26a69a\",\"response\":\"Resource deleted successfully.\"}");
-            } catch (JDOObjectNotFoundException e){
-                System.err.println("Exception catched -> " + e.getMessage());
+        try{
+
+            if (AccessControllerView.checkPermission(request.getSession().getAttribute("userID").toString(),request.getRequestURI())){
+
+                PersistenceManager pm = controller.PMF.get().getPersistenceManager();
+
+                try {
+                    Key key = KeyFactory.stringToKey(request.getParameter("key"));
+                    try{
+                        pm.deletePersistent(pm.getObjectById(Resource.class, key));
+                        request.getSession().setAttribute("serverResponse","{\"color\": \"#26a69a\",\"response\":\"Recurso eliminado con éxito.\"}");
+                    } catch (JDOObjectNotFoundException e){
+                        System.err.println("Exception catched -> " + e.getMessage());
+                    }
+
+
+                } catch (NullPointerException e){
+                    System.err.println("Exception captured -> " + e.getMessage());
+                }
+
+                response.sendRedirect("/e/resources");
+
+            } else {
+                request.getSession().setAttribute("serverResponse","{\"color\": \"red\",\"response\":\"No tienes permiso para eliminar un Recurso.\"}");
+                response.sendRedirect("/e/resources");
             }
 
-
         } catch (NullPointerException e){
-            System.err.println("Exception captured -> " + e.getMessage());
+            response.sendRedirect("/");
         }
-
-        response.sendRedirect("/e/resources");
         
     }
